backend_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
anthropic==0.28.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.23.0
pytest-mock==3.12.0
mongomock-motor==0.0.36
